Accept empty payloads for embed attributes
This does not set the attributes (they are set to None and therefore
not present in the output).
However, this allows for an empty payload to be parsed out from the
script should this be desired and/or necessary (e.g. a title may end
up empty because of some dynamic conditions).