Skip duplicate spaces in ReactBlock payload
Previously this could lead to false positive "limit reached" messages
if there were too many spaces between emoji.

This is not an issue with the zero-depth splitter itself, so I added
a test to maintain this behaviour. The ReactBlock just filters the
empty split output elements itself. Tests for this behaviour have been
added for `react` and `reactu` as well.