Impose "zero-depth" restrictions on EmbedBlock
Previously, the payload would be parsed before passing it on to the
appropriate attribute handler.
This caused issues for a script where usernames were included in embed
footers, specifically when users with "|" in their names would get
interpreted, then split once by the footer setter, causing the
`icon_url` attribute to contain the rest of the username as well as
the icon URL, leading to errors from Discord's end for malformed URLs.

The best way around this is to prevent the "injection" by imposing the
zero-depth restriction on EmbedBlock use (since this restriction was
designed to avoid injections in the conditional block group).